More work on SPARQL UPDATE support.  Refactored to derive an operator to REMOVE statements from the same base class as the operator to ADD statements.  Added support for reporting mutation counts on relations through the bop statistics and extended that support through the painting of the EXPLAIN page.  Added the ability to identify vocabulary IVs via isVocabulary().  LexiconRelation#addTerms() now reports the #of terms added (or the #of terms which would have been added if invoked with readOnly=true).
@see https://sourceforge.net/apps/trac/bigdata/ticket/448 (SPARQL 1.1 UPDATE)

git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6155 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/java/com/bigdata/rdf/lexicon/WriteTaskStats.java
@@ -14,9 +14,10 @@ import com.bigdata.counters.CAT;
 public class WriteTaskStats {
     
     /**
-     * The #of distinct terms lacking a pre-assigned term identifier in [a].
+     * The #of distinct terms lacking a pre-assigned term identifier. If writes
+     * were permitted, then this is also the #of terms written onto the index.
      */
-    final AtomicInteger ndistinct = new AtomicInteger();
+    final AtomicLong ndistinct = new AtomicLong();
 
     /** time to convert unicode terms to byte[] sort keys. */
     final CAT keyGenTime = new CAT();
